+++ scenarios/simple/bt_nodes.py
@@ -1,7 +1,6 @@
-import math
 import random
-from modules.base_bt_nodes import BTNodeList, Status, Node, Sequence, Fallback, ReactiveSequence, ReactiveFallback, SyncAction, GatherLocalInfo, AssignTask
-from modules.base_bt_nodes import _IsTaskCompleted, _IsArrivedAtTask, _MoveToTask, _ExecuteTaskWhileFollowing, _ExploreArea
+from modules.base_bt_nodes import BTNodeList, Status, Node, Sequence, Fallback, ReactiveSequence, ReactiveFallback, SyncAction, SyncCondition, GatherLocalInfo, AssignTask
+
 # BT Node List
 CUSTOM_ACTION_NODES = [
     'MoveToTarget',
@@ -14,64 +13,95 @@ CUSTOM_CONDITION_NODES = [
     'IsArrivedAtTarget',
 ]
 
-# BT Node List
 BTNodeList.ACTION_NODES.extend(CUSTOM_ACTION_NODES)
 BTNodeList.CONDITION_NODES.extend(CUSTOM_CONDITION_NODES)
 
 
-# Scenario-specific Action/Condition Nodes
+# Scenario-specific config
 from modules.utils import config
 target_arrive_threshold = config['tasks']['threshold_done_by_arrival']
 task_locations = config['tasks']['locations']
 sampling_freq = config['simulation']['sampling_freq']
-sampling_time = 1.0 / sampling_freq  # in seconds
+sampling_time = 1.0 / sampling_freq
 agent_max_random_movement_duration = config.get('agents', {}).get('random_exploration_duration', None)
 
 
-class IsTaskCompleted(_IsTaskCompleted): 
+class IsTaskCompleted(SyncCondition):
     def __init__(self, name, agent):
-        super().__init__(name, agent)   
+        super().__init__(name, self._update)
 
-    def _update(self, agent, blackboard): 
-        result = super()._update(agent, blackboard, task_id_key='assigned_task_id')
-        if result is Status.SUCCESS:
+    def _update(self, agent, blackboard):
+        assigned_task_id = blackboard.get('assigned_task_id')
+        if assigned_task_id is None:
+            return Status.RUNNING
+
+        task = agent.tasks_info[assigned_task_id]
+        if task.completed is True:
             blackboard['assigned_task_id'] = None
-        return result
+            return Status.SUCCESS
+        return Status.FAILURE
 
 
-class IsArrivedAtTarget(_IsArrivedAtTask): 
+class IsArrivedAtTarget(SyncCondition):
     def __init__(self, name, agent):
-        super().__init__(name, agent)   
+        super().__init__(name, self._update)
 
-    def _update(self, agent, blackboard): 
-        result = super()._update(agent, blackboard, task_id_key='assigned_task_id', arrive_threshold=target_arrive_threshold)
-        if result is Status.SUCCESS:
-            pass
-        return result
-    
-    
-class MoveToTarget(_MoveToTask): 
+    def _update(self, agent, blackboard):
+        assigned_task_id = blackboard.get('assigned_task_id')
+        if assigned_task_id is None:
+            raise ValueError(f"[{self.name}] Error: No assigned_task_id found in the blackboard!")
+
+        distance = (agent.tasks_info[assigned_task_id].position - agent.position).length()
+        if distance < agent.tasks_info[assigned_task_id].radius + target_arrive_threshold:
+            return Status.SUCCESS
+        return Status.FAILURE
+
+
+class MoveToTarget(SyncAction):
     def __init__(self, name, agent):
-        super().__init__(name, agent)   
+        super().__init__(name, self._update)
 
-    def _update(self, agent, blackboard): 
-        result = super()._update(agent, blackboard, task_id_key='assigned_task_id')
-        return result
-        
-class ExecuteTask(_ExecuteTaskWhileFollowing): 
+    def _update(self, agent, blackboard):
+        assigned_task_id = blackboard.get('assigned_task_id')
+        if assigned_task_id is None:
+            raise ValueError(f"[{self.name}] Error: No assigned_task_id found in the blackboard!")
+
+        agent.follow(agent.tasks_info[assigned_task_id].position)
+        return Status.RUNNING
+
+
+class ExecuteTask(SyncAction):
     def __init__(self, name, agent):
-        super().__init__(name, agent)   
+        super().__init__(name, self._update)
 
-    def _update(self, agent, blackboard): 
-        result = super()._update(agent, blackboard, task_id_key='assigned_task_id')
-        return result
-                
+    def _update(self, agent, blackboard):
+        assigned_task_id = blackboard.get('assigned_task_id')
+        if assigned_task_id is None:
+            raise ValueError(f"[{self.name}] Error: No assigned_task_id found in the blackboard!")
 
-class Explore(_ExploreArea): 
+        agent.tasks_info[assigned_task_id].reduce_amount(agent.work_rate)
+        agent.update_task_amount_done(agent.work_rate)
+        agent.follow(agent.tasks_info[assigned_task_id].position)
+        return Status.RUNNING
+
+
+class Explore(SyncAction):
     def __init__(self, name, agent):
-        super().__init__(name, agent)   
+        super().__init__(name, self._update)
+        self.random_move_time = float('inf')
+        self.random_waypoint = (0, 0)
 
-    def _update(self, agent, blackboard): 
-        result = super()._update(agent, blackboard, agent_max_random_movement_duration=agent_max_random_movement_duration, exploration_area=task_locations, sampling_time=sampling_time)
-        return result
-        
+    def _update(self, agent, blackboard):
+        if self.random_move_time > agent_max_random_movement_duration:
+            self.random_waypoint = (
+                random.randint(task_locations['x_min'], task_locations['x_max']),
+                random.randint(task_locations['y_min'], task_locations['y_max'])
+            )
+            self.random_move_time = 0
+
+        self.random_move_time += sampling_time
+        agent.follow(self.random_waypoint)
+        return Status.RUNNING
+
+    def halt(self):
+        self.random_move_time = float('inf')
